fix transparency, sort transparent

+++ src/simulation/chunk.rs
@@ -4,8 +4,8 @@ use cgmath::Vector3;
 
 #[derive(Debug, Clone)]
 pub struct Chunk {
-    pub id: u64,
-    pub position: Vector3<i64>,
+    pub id: u32,
+    pub position: Vector3<i32>,
     pub modified: bool,
     pub blocks: Box<[Block; CHUNK_VOLUME as usize]>,
 }
